requirements.txt
dnspython>=2.0.0
python-whois>=0.7.3